Fix isGoodTrack integer-division bug in the fitMaxRatio hit-fraction check
Both call sites (line ~1325, the event-loop track gate, and line
~1389, the goodTrack bool used for DCA-histogram gating) passed
track->nHits()/track->nHitsPoss() -- integer division, since both are
Int_t -- into isGoodTrack()'s double a_nHitsOverNPoss parameter. The
division truncated to 0 for virtually every track (nHits < nHitsPoss
in the overwhelming majority of cases) or 1 only when they were
exactly equal, so the fitMaxRatio>=0.5001 check was effectively
'reject unless nHitsFit==nHitsMax exactly' instead of the intended
>=50% ratio -- silently disabling the split/broken-track rejection
this cut exists for, for the entire history of this analysis.

Fixed by casting to double before dividing at both call sites. This is
also why eta_nHitsFit and eta_nHitsMax looked identical: every track
that survived to be filled into either histogram already satisfied
nHitsFit==nHitsMax by construction of the bug.

Does not touch CutClass.cxx -- the other two isGoodTrack() overloads
(StPicoTrack*-based, and the 6-arg raw-nHitsMax version) already did
this division correctly with double casts; the bug was only in how
PicoBinner.cxx called the 5-arg pre-computed-ratio overload.

+++ source/PicoBinner.cxx
@@ -1322,7 +1322,12 @@ void PicoBinner(string    a_filelist,
 
     for (int trackIndex = 0; trackIndex < (int)dst->numberOfTracks(); trackIndex++){
       StPicoTrack* track = dst->track(trackIndex);
-      if(!a_cutClass->isGoodTrack(track->nHitsFit(), track->nHits()/track->nHitsPoss(),track->nHitsDedx(), track->gDCA(eventVertex).Mag(), phi)) continue;
+      // Bug fix 2026-07-11: nHits()/nHitsPoss() was INTEGER division (both are Int_t),
+      // truncating to 0 unless a track used literally every possible padrow -- this made
+      // the fitMaxRatio cut inside isGoodTrack() effectively "reject unless nHitsFit ==
+      // nHitsMax exactly" instead of the intended >=50% ratio, silently defeating the
+      // split/broken-track rejection this cut exists for. Cast to double BEFORE dividing.
+      if(!a_cutClass->isGoodTrack(track->nHitsFit(), ((double)track->nHits())/((double)track->nHitsPoss()),track->nHitsDedx(), track->gDCA(eventVertex).Mag(), phi)) continue;
       double eta = track->pMom().Eta();
       if(eta >= -1.0 && eta <= 0.0) refMultHighVar++;
       if(eta >= 0.0 && eta <= 1.0) refMultLowVar++;
@@ -1381,7 +1386,9 @@ void PicoBinner(string    a_filelist,
       // further down this loop -- computed once here and reused everywhere below,
       // identical values, just without the redundant recomputation.
       double dca = track->gDCA(eventVertex).Mag();
-      bool goodTrack = a_cutClass->isGoodTrack(track->nHitsFit(), track->nHits()/track->nHitsPoss(), track->nHitsDedx(), dca, phi);
+      // Same integer-division fix as the isGoodTrack() call above -- cast to double
+      // before dividing, see that call site's comment for the full explanation.
+      bool goodTrack = a_cutClass->isGoodTrack(track->nHitsFit(), ((double)track->nHits())/((double)track->nHitsPoss()), track->nHitsDedx(), dca, phi);
       //FILL DCA for Proton and AntiProton
       if(doProtonDCA && (a_partIndex < 0 || a_partIndex == 2) && goodTrack){
         double rapidity = PhysMath::rapidity(particleInfo->GetParticleMass(2), pT, pZ);
